break wifi_mgr.c up to split the main function to wifi_mgr_main.c, necessary for building tests without breaking the linker
